Add grid view and update command, option_type module and configuration

+++ virtTrinity/option_type.py
@@ -32,13 +32,28 @@ class RandomNumber(RandomString):
         self.line = str(random.randint(-2, 100))
 
 
+class RandomNonNegative(RandomNumber):
+    def __init__(self, opt_args):
+        self.line = str(random.randint(0, 100))
+
+
+class RandomHugePageSize(RandomNumber):
+    def __init__(self, opt_args):
+        self.line = "2048"
+
+
+class RandomCellNumber(RandomNumber):
+    def __init__(self, opt_args):
+        self.line = "0"
+
+
 class RandomRebootMode(RandomString):
     def __init__(self, opt_args):
         self.line = random.choice([
             "acpi", "agent", "initctl", "signal", "paravirt"])
 
 
-class RandomFd(RandomString):
+class RandomFd(RandomNumber):
     pass
 
 
@@ -87,6 +102,56 @@ class RandomDomain(RandomString):
         ]
         if vm_names:
             self.line = random.choice(vm_names)
+            opt_args['domain'] = self.line
+        else:
+            opt_args['skip'] = True
+
+
+class RandomActiveDomain(RandomDomain):
+    def __init__(self, opt_args):
+        self.line = 'virt-trinity-1'
+        vm_names = [
+            name for name in
+            utils.run('virsh list --name').stdout.splitlines()
+            if name.startswith('virt-trinity-')
+        ]
+        if vm_names:
+            self.line = random.choice(vm_names)
+            opt_args['domain'] = self.line
+        else:
+            opt_args['skip'] = True
+
+
+class RandomInactiveDomain(RandomDomain):
+    def __init__(self, opt_args):
+        self.line = 'virt-trinity-1'
+        vm_names = [
+            name for name in
+            utils.run('virsh list --inactive --name').stdout.splitlines()
+            if name.startswith('virt-trinity-')
+        ]
+        if vm_names:
+            self.line = random.choice(vm_names)
+            opt_args['domain'] = self.line
+        else:
+            opt_args['skip'] = True
+
+
+class RandomUnexistDomain(RandomDomain):
+    def __init__(self, opt_args):
+        self.line = 'virt-trinity-1'
+        all_names = ['virt-trinity-' + str(i) for i in xrange(1, 10)]
+        vm_names = [
+            name for name in
+            utils.run('virsh list --all --name').stdout.splitlines()
+            if name.startswith('virt-trinity-')
+        ]
+        avail_names = list(set(all_names) - set(vm_names))
+        if avail_names:
+            self.line = random.choice(avail_names)
+            opt_args['domain'] = self.line
+        else:
+            opt_args['skip'] = True
 
 
 class RandomFile(RandomString):
@@ -100,19 +165,33 @@ class RandomFile(RandomString):
             xml_file.write(self.content)
 
     def post(self):
-        os.remove(self.path)
+        if os.path.exists(self.path):
+            os.remove(self.path)
 
 
 class RandomVmXml(RandomFile):
     def __init__(self, opt_args):
         self.line = 'virt-trinity-vm.xml'
         self.path = self.line
-        self.content = """
-            <domain type='kvm'>
-              <name>virt-trinity-%s</name>
-              <memory>100000</memory>
-              <os><type>hvm</type></os>
-            </domain>""" % random.randint(1, 9)
+
+        vm_name = 'virt-trinity-1'
+        all_names = ['virt-trinity-' + str(i) for i in xrange(1, 10)]
+        vm_names = [
+            name for name in
+            utils.run('virsh list --all --name').stdout.splitlines()
+            if name.startswith('virt-trinity-')
+        ]
+        avail_names = list(set(all_names) - set(vm_names))
+        if avail_names:
+            vm_name = random.choice(avail_names)
+            self.content = """
+                <domain type='kvm'>
+                  <name>%s</name>
+                  <memory>100000</memory>
+                  <os><type>hvm</type></os>
+                </domain>""" % vm_name
+        else:
+            opt_args['skip'] = True
 
 
 class RandomDeviceXml(RandomFile):
@@ -172,17 +251,131 @@ class RandomVolXml(RandomFile):
         """ % (self.number, self.number)
 
 
-class RandomExistingDeviceXml(RandomDeviceXml):
+class RandomExistDisk(RandomString):
+    def __init__(self, opt_args):
+        self.line = "vda"
+        if 'domain' in opt_args:
+            text = utils.run(
+                'virsh -q domblklist %s' % opt_args['domain']).stdout
+            disks = [l.split()[0] for l in text.strip().splitlines()]
+            if disks:
+                self.line = random.choice(disks)
+            else:
+                opt_args['skip'] = True
+        else:
+            opt_args['skip'] = True
+
+
+class RandomDiskCountString(RandomString):
+    def __init__(self, opt_args):
+        self.line = "vda,1"
+        if 'domain' in opt_args:
+            text = utils.run(
+                'virsh -q domblklist %s' % opt_args['domain']).stdout
+            disks = [l.split()[0] for l in text.strip().splitlines()]
+            if disks:
+                disk_strs = []
+                for disk in disks:
+                    disk_strs.append(disk)
+                    disk_strs.append(str(random.randint(-2, 100)))
+                if disk_strs:
+                    self.line = ",".join(disk_strs)
+                else:
+                    opt_args['skip'] = True
+            else:
+                opt_args['skip'] = True
+        else:
+            opt_args['skip'] = True
+
+
+class RandomDiskSourceType(RandomString):
+    def __init__(self, opt_args):
+        self.line = random.choice(['file', 'block'])
+
+
+class RandomDiskAddress(RandomString):
+    def __init__(self, opt_args):
+        bus_type = random.choice(['pci', 'scsi', 'ide'])
+        n = 3
+        if bus_type == 'pci':
+            n = 4
+        addr_list = [str(random.randint(1, 9)) for i in xrange(n)]
+        self.line = '%s:%s' % (bus_type, '.'.join(addr_list))
+
+
+class RandomDiskType(RandomString):
+    def __init__(self, opt_args):
+        self.line = random.choice(['lun', 'cdrom', 'floppy'])
+
+
+class RandomDiskSerial(RandomString):
+    def __init__(self, opt_args):
+        chars = string.ascii_letters + string.digits + '_-+.'
+        length = random.randint(5, 10)
+        self.line = ''.join(random.choice(chars) for _ in xrange(length))
+
+
+class RandomDiskName(RandomString):
+    def __init__(self, opt_args):
+        prefix = random.choice(['hd', 'sd', 'vd', 'xvd', 'ubd'])
+        postfix = random.choice(string.ascii_letters)
+        self.line = prefix + postfix
+
+
+class RandomDiskBus(RandomString):
+    def __init__(self, opt_args):
+        self.line = random.choice(['ide', 'scsi', 'virtio', 'xen', 'usb',
+                                   'sata', 'sd'])
+
+
+class RandomDiskDriver(RandomString):
+    def __init__(self, opt_args):
+        self.line = random.choice(['file', 'tap', 'phy', 'qemu'])
+
+
+class RandomDiskSubDriver(RandomString):
+    def __init__(self, opt_args):
+        self.line = random.choice(['aio', 'raw', 'qcow2'])
+
+
+class RandomDiskIothread(RandomNumber):
+    pass
+
+
+class RandomDiskCache(RandomString):
+    def __init__(self, opt_args):
+        self.line = random.choice(['default', 'none', 'writethrough',
+                                   'writeback', 'directsync', 'unsafe'])
+
+
+class RandomDiskMode(RandomString):
+    def __init__(self, opt_args):
+        self.line = random.choice(['readonly', 'shareable'])
+
+
+class RandomWwn(RandomNumber):
+    def __init__(self, opt_args):
+        self.line = ''.join([str(random.choice(string.hexdigits))
+                             for i in xrange(16)])
+
+
+class RandomExistDeviceXml(RandomDeviceXml):
     def __init__(self, opt_args):
         self.path = 'virt-trinity-device.xml'
         self.line = self.path
         self.content = '<>'
-        xml = utils.run('virsh dumpxml virt-trinity-vm1').stdout
-        if xml.strip():
-            root = ElementTree.fromstring(xml)
-            iface_xmls = root.findall("./devices/interface")
-            if iface_xmls:
-                self.content = ElementTree.tostring(random.choice(iface_xmls))
+        if 'domain' in opt_args:
+            xml = utils.run('virsh dumpxml %s' % opt_args['domain']).stdout
+            if xml.strip():
+                root = ElementTree.fromstring(xml)
+                iface_xmls = root.findall("./devices/interface")
+                if iface_xmls:
+                    self.content = ElementTree.tostring(
+                        random.choice(iface_xmls))
+                else:
+                    opt_args['skip'] = True
+        else:
+            opt_args['skip'] = True
 
 
 def parse_type(type_name, opt_args):
